+++ PGPatcher/include/PGConfig.hpp
@@ -133,7 +133,12 @@ public:
 
         // Global Patchers.
         struct GlobalPatcher {
-            bool operator==([[maybe_unused]] const GlobalPatcher& other) const { return true; }
+            bool isRecalculateBoundsEnabled = false;
+
+            bool operator==(const GlobalPatcher& other) const
+            {
+                return isRecalculateBoundsEnabled == other.isRecalculateBoundsEnabled;
+            }
         } globalPatcher;
 
         bool operator==(const PGParams& other) const